Reorganise code in preparation for HashedPassword class
Two changes here:
1. extracts most of the Password code to PasswordBase, which will be used for HashedPassword
2. moves the hash_password method to Pwned, so it is exposed clearly and publicly if anyone wants to hash their password before checking whether it's been pwned
